% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{somedat}
\alias{somedat}
\title{somedat - fake data set as an example}
\source{
Made-up by internal function \code{.dataCreation()}
}
\usage{
data(somedat)
}
\description{
This data set contains a columns of data, a time-trend variable
foo and a noise variable bar
}
\author{
Dirk Eddelbuettel
}
\keyword{datasets}

